+++ apps/docs/pages/docs/api-reference/configuration/_meta.js
@@ -1,0 +1,5 @@
+const menu = {
+  config: {},
+};
+
+export default menu;